Added changes to ensureDHkeysize after some discussions
